camera and background adjustments
I changed the positioning of the camera, the character and the dance floor, so that it is easier to use.

No weird rotations or perspectives.

Everything is aligned at the X axis, at 0.

To offset the dancer to the left side of the screen, I used the camera offset on x axis - this keeps the character alight with the vanishing point of the camera.