Bug: pause menu was starting the game while keeping the menu on - fixed
before starting the game, I was able to pause it. If I unpause it without pressing any other key, the music and notes would start rolling, but the press any key to start UI would keep existing.

this was also dependent on the exclusion of the space key to start the game - witch was included to prevent conflicts with the pause menu.

Now if you haven't start the level, even if you press space, you won't open the pause menu, you will first start the game, then, you can press again the space key to pause it. This way, you can really press any key to start the level, and when the level is started, you can pause it.